+++ src/ItaasSolution.Api/Domain/Entities/Log.cs
@@ -6,10 +6,10 @@ namespace ItaasSolution.Api.Domain.Entities
     {
         public long Id { get; set; }
         public string HtttpMethod { get; set; }
-        public long StatusCode { get; set; }
+        public int StatusCode { get; set; }
         public string UriPath { get; set; }
         public decimal TimeTaken { get; set; }
-        public long ResponseSize { get; set; }
+        public int ResponseSize { get; set; }
         public string CacheStatus { get; set; }
         public DateTime DateCreation { get; set; }
     }
